+++ cfdtools/vtk/__init__.py
@@ -1,1 +1,1 @@
-from cfdtools.vtk._vtk import vtkMesh
+from cfdtools.vtk._vtk import vtkMesh, vtkList