+++ CHANGELOG.md
@@ -16,6 +16,15 @@ pip show cfdtools
 
 ## Versions
 
+### [0.5.0](https://pypi.org/project/cfdtools/) (2023-04-05)
+
+#### new
+
+- legendre polynomial extrapolation (development)
+- DataSet and DataSetList classes
+- read list of vtk files to DataSetList
+- dump DataSetList as hdf5 file
+
 ### [0.4.2](https://pypi.org/project/cfdtools/) (2023-04-05)
 
 #### new